# openapi.yaml
openapi: 3.0.0
info:
  title: API de Asignación P3 - Gabriel Ochoa
  version: '1.0.0'
  description: Documentación OpenAPI generada para el proyecto P3. Contiene todas las rutas públicas y protegidas.
  contact:
    name: Gabriel Andres Ochoa Padron
    email: 31903665
servers:
  - url: '/' 
    description: Servidor local (ruta base)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        nombreCompleto:
          type: string
        email:
          type: string
    AuthResponse:
      type: object
      properties:
        status:
          type: string
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        stock:
          type: integer
        sku:
          type: string
        brand:
          type: string
        slug:
          type: string
        categoryId:
          type: integer
        category:
          $ref: '#/components/schemas/Category'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
...
paths:
...
  /auth/register:
    post:
      summary: Registrar nuevo usuario
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombreCompleto, email, password]
              properties:
                nombreCompleto:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Error en datos de entrada
  /auth/login:
    post:
      summary: Iniciar sesión
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciales inválidas
  /users:
    get:
      summary: Obtener lista de usuarios (protegido)
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          description: No autorizado
    post:
      summary: Crear usuario (protegido)
      tags:
        - Users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: Usuario creado
...
  /products:
    get:
      summary: Lista pública de productos con filtros
      tags:
        - Public - Products
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: category
          schema:
            type: integer
        - in: query
          name: tags
          schema:
            type: string
          description: Comma-separated tag IDs, e.g. 1,2,3
        - in: query
          name: price_min
          schema:
            type: number
        - in: query
          name: price_max
          schema:
            type: number
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: brand
          schema:
            type: string
      responses:
        '200':
          description: Lista paginada de productos
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
    post:
      summary: Crear producto (protegido)
      tags:
        - Admin - Products
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                stock:
                  type: integer
                sku:
                  type: string
                brand:
                  type: string
                categoryId:
                  type: integer
                tags:
                  type: array
                  items:
                    type: integer
      responses:
        '201':
          description: Producto creado
...
  /p/{id}-{slug}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Obtener producto público por id y slug (self-healing)
      tags:
        - Public - Products
      responses:
        '200':
          description: Producto retornado correctamente
        '301':
          description: Redirección a la URL canónica si el slug difiere
        '404':
          description: Producto no encontrado
  /categories:
    get:
      summary: Listar categorías (protegido)
      tags:
        - Admin - Categories
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de categorías
    post:
      summary: Crear categoría (protegido)
      tags:
        - Admin - Categories
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '201':
          description: Categoría creada
...
  /tags:
    get:
      summary: Listar tags (protegido)
      tags:
        - Admin - Categories
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de tags
    post:
      summary: Crear tag (protegido)
      tags:
        - Admin - Categories
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
      responses:
        '201':
          description: Tag creado
...